LayoutTests/userscripts/user-script-plugin-document.html

<html>
<head>
<script>
if (window.testRunner) {
    window.testRunner.dumpAsText();
    window.testRunner.addUserScript("console.log('Start Script: ' + document.location.href)", true, true);
    window.testRunner.addUserScript("console.log('End Script: ' + document.location.href)", false, true);
}
</script>
</head>
<body>
<iframe src="data:application/x-webkit-test-netscape,"></iframe>
</body>
</html>
